+++ db/updates/6.6-7.sql
@@ -1,0 +1,1 @@
+DROP TABLE `pv_online`;