+++ func/pvOnline.php
@@ -15,17 +15,6 @@ function pvOnlineStats($api, $s){
 	
 }
 
-function pvOnlineValAcc($api, $apiUser, $apiPass, $ver){
-	$jAPI = $api.'?username='.$apiUser.'&password='.$apiPass.'&login=1&ver='.$ver;
-	$ch = curl_init($jAPI);
-	curl_setopt($ch, CURLOPT_RETURNTRANSFER, true);
-	curl_setopt($ch, CURLOPT_HTTPHEADER, array('Content-Type:application/json'));
-
-	$response = json_decode(curl_exec($ch),true);
-	curl_close($ch);
-	return $response['auth'];
-}
-
 function pvUploadData($pvOnlineAPI, $data){
 	$ch = curl_init($pvOnlineAPI);
 	curl_setopt($ch, CURLOPT_RETURNTRANSFER, true);